fix requests change
