+++ app/assets/javascripts/bp_metadata.js.erb
@@ -1,0 +1,5 @@
+jQuery("#search_metadata").chosen({
+    search_contains: true,
+    width: "432px"
+});
+
