+++ app/assets/stylesheets/application.css.scss.erb
@@ -40,7 +40,7 @@
 @import "ontolobridge";
 @import "fair_assement";
 @import "instances_table";
-
+@import "metadata";
 /* Bootstrap and Font Awesome */
 @import "bootstrap";
 @import "bootstrap_overrides";
