+++ app/assets/stylesheets/metadata.scss
@@ -1,0 +1,7 @@
+#search_metadata_chosen .chosen-container .chosen-container-multi {
+    width: 432px;
+}
+
+.horiz{
+    display: inline-block; 
+}
