+++ app/helpers/ontology_metadata_curator_helper.rb
@@ -1,0 +1,14 @@
+module OntologyMetadataCuratorHelper
+  def metadata_for_select
+    get_metadata
+    return @metadata_for_select
+  end 
+
+  def get_metadata
+    @metadata_for_select = []
+    submission_metadata.each do |data|
+      @metadata_for_select << data["attribute"]
+    end
+    @metadata_for_select.sort! 
+  end   
+end
