--- app/views/ontologies_metadata/_form_edit.html.haml
@@ -1,30 +1,0 @@
-:css
-  .horiz{
-    display: inline-block; 
-  }
-
-= turbo_frame_tag "editmodal" do
-  -#= tag.div :div, data: { controller: "modal", modal_target: "modal", action: "click@window->modal#closeBackground", style: "padding: 30px;"} do   
-  .modal.fade{"aria-hidden" => "true", "aria-labelledby" => "exampleModalLabel", "data-controller" => "modal" ,role: "dialog", tabindex: "-1"}
-    .modal-dialog{role: "document"}
-      .modal-content
-        .modal-header
-          %h5.modal-title Modal title
-          %button.close{"aria-label" => "Close", "data-bs-dismiss" => "modal", type: "button"}
-            %span{"aria-hidden" => "true"} ×
-        .modal-body
-          = form_with(url: "#", method: :post) do |form|
-            %h3 
-              Selected Ontologies
-              %br
-              -@selected_ontologies_to_edit.each do |ontology|
-                %div.horiz= ontology 
-            %table.form
-              - @selected_metadata_to_edit.each do |meta|
-                %tr
-                  %th= meta
-                  %td.top= form.text_field :submission, value: @submission.instance_values[meta.to_s]
-        .modal-footer
-          %button.btn.btn-secondary{"data-bs-dismiss" => "modal", type: "button"} Close
-          %button.btn.btn-primary{type: "button"} Save changes
-
