--- app/views/ontologies_metadata/result.turbo_stream.haml
@@ -1,2 +1,0 @@
-= turbo_stream.append  "choices", partial: "result"
-
