+++ app/views/ontologies_metadata_curator/_form_edit.html.haml
@@ -1,0 +1,16 @@
+= turbo_frame_tag "editmodal" do
+  -#= tag.div :div, data: { controller: "modal", modal_target: "modal", action: "click@window->modal#closeBackground", style: "padding: 30px;"} do   
+  .modal.fade{"aria-hidden" => "true", "aria-labelledby" => "exampleModalLabel", "data-controller" => "modal" ,role: "dialog", tabindex: "-1"}
+    .modal-dialog{role: "document"}
+      .modal-content
+        .modal-header
+          %h5.modal-title Modal title
+          %button.close{"aria-label" => "Close", "data-bs-dismiss" => "modal", type: "button"}
+            %span{"aria-hidden" => "true"} ×
+        .modal-body
+          = form_for(:submission, url: "/update" ,html: {method: :put}) do |f|
+            = render partial: 'form', locals: { f: f } 
+            .modal-footer
+              %button.btn.btn-secondary{"data-bs-dismiss" => "modal", type: "button"} Close
+              =f.submit "Save changes", :class => "blueButton"
+
