+++ app/views/ontologies_metadata_curator/_result.html.haml
@@ -1,5 +1,5 @@
 = turbo_frame_tag "metadata" do
-  = form_with(url: "/edit", method: :post, local: true, data: { turbo: true} ,form: {data: {turbo_frame: 'editmodal'}}) do |form|  
+  = form_with(url: "/edit", method: :post, local: true, data: { turbo: true}) do |form|  
     %table.table{:cellpadding => "0", :cellspacing => "0"}
       %col
       %colgroup{:span=>"1"}
