+++ app/views/submissions/edit.html.haml
@@ -1,6 +1,3 @@
-- @title = "Edit ontology submission"
-
-
 %div.container.py-4.py-md-5
   %h3.text-center.mb-4
     Edit submission information
